+++ de.xwic.appkit.webbase/src/de/xwic/appkit/webbase/table/RowData.java
@@ -7,12 +7,12 @@
  *
  * 		http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software 
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and 
+ * See the License for the specific language governing permissions and
  * limitations under the License.
- *  
+ *
  *******************************************************************************/
 
 package de.xwic.appkit.webbase.table;
@@ -26,32 +26,32 @@ import de.xwic.appkit.core.dao.IEntity;
 public class RowData {
 
 	private long entityId = 0;
-	
+
 	private Object[] array = null;
 	private IEntity entity = null;
 	private Object object = null;
 
 	private final EntityTableModel model;
-	
-	
+
+
 	/**
-	 * @param o 
+	 * @param o
 	 * @param o
 	 */
 	public RowData(EntityTableModel model, Object data) {
-		
+
 		this.model = model;
 		object = data;
 		if (data.getClass().isArray()) {
 			array = (Object[])data;
-			entityId = (Integer)array[0];
+			entityId = (Long)array[0];
 		} else if (data instanceof IEntity) {
 			entity = (IEntity)data;
 			entityId = entity.getId();
 		}
-		
+
 	}
-	
+
 	/**
 	 * Returns true if the data is available as array.
 	 * @return
@@ -59,7 +59,7 @@ public class RowData {
 	public boolean isArray() {
 		return array != null;
 	}
-	
+
 	/**
 	 * Returns the data of the given property.
 	 * @param propertyId
